pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
youtube-transcript-api = "^0.6.2"
docx = "^0.2.4"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
discord-py = "^2.3.2"
google-generativeai = "^0.3.2"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"
giphy-client = "^1.0.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"